requirements.txt
faiss-cpu==1.10.0
flask==3.1.0
twilio==9.3.6
numpy==2.2.3
sentence-transformers==3.2.1